requirements.txt
# Core Data Science
numpy>=1.24.0,<2.0.0
pandas>=2.2.0
scipy>=1.11.0

# Data Sources
yfinance>=0.2.28

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging & Progress
loguru>=0.7.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0